fix fork ss, update files based on review
