+++ CPAC/seg_preproc/seg_preproc.py
@@ -1070,3 +1070,231 @@ def create_seg_preproc_antsJointLabel_method(wf_name='seg_preproc_templated_base
                     outputNode, 'wm_mask')
 
     return preproc
+
+
+def connect_anat_segmentation(workflow, strat_list, c, strat_name=None):
+    
+    """
+    Segmentation Preprocessing Workflow
+    
+    Parameters
+    ----------
+        workflow : workflow
+            main preprocessing workflow
+        strat_list : list
+            list of strategies
+        c : configuration
+            pipeline configuration
+        strat_name : str
+            name of strategy
+
+    Returns
+    -------
+        workflow : workflow
+            updated main preprocessing workflow
+        strat_list : list
+            list of updated strategies
+    """
+    
+    from CPAC.seg_preproc.seg_preproc import (
+        create_seg_preproc,
+        create_seg_preproc_template_based
+    )
+
+    new_strat_list = []
+
+    if 1 in c.runSegmentationPreprocessing:
+
+        for num_strat, strat in enumerate(strat_list):
+
+            nodes = strat.get_nodes_names()
+
+            seg_preproc = None
+
+            if not any(o in c.seg_use_threshold for o in ["FSL-FAST Thresholding", "Customized Thresholding"]):
+                err = '\n\n[!] C-PAC says: Your segmentation thresholding options ' \
+                    'setting does not include either \'FSL-FAST Thresholding\' or \'Customized Thresholding\'.\n\n' \
+                    'Options you provided:\nseg_use_threshold: {0}' \
+                    '\n\n'.format(str(c.seg_use_threshold))
+                raise Exception(err)
+
+            if strat.get('registration_method') == 'FSL':
+                use_ants = False
+            elif strat.get('registration_method') == 'ANTS':
+                use_ants = True
+
+            if strat_name != None:
+                seg_preproc_wf_name = f'seg_preproc_{strat_name}_{num_strat}'
+            else:
+                seg_preproc_wf_name = f'seg_preproc_{num_strat}'
+
+            seg_preproc = create_seg_preproc(use_ants=use_ants,
+                                             use_priors=c.seg_use_priors,
+                                             use_threshold=c.seg_use_threshold,
+                                             csf_use_erosion=c.seg_csf_use_erosion,
+                                             wm_use_erosion=c.seg_wm_use_erosion,
+                                             gm_use_erosion=c.seg_gm_use_erosion,
+                                             wf_name=seg_preproc_wf_name)
+
+            seg_preproc.inputs.csf_threshold.csf_threshold=c.seg_CSF_threshold_value
+            seg_preproc.inputs.wm_threshold.wm_threshold=c.seg_WM_threshold_value
+            seg_preproc.inputs.gm_threshold.gm_threshold=c.seg_GM_threshold_value
+
+            seg_preproc.inputs.csf_erosion_prop.csf_erosion_prop=c.csf_erosion_prop
+            seg_preproc.inputs.wm_erosion_prop.wm_erosion_prop=c.wm_erosion_prop
+            seg_preproc.inputs.gm_erosion_prop.gm_erosion_prop=c.gm_erosion_prop
+
+            seg_preproc.inputs.csf_mask_erosion_mm.csf_mask_erosion_mm=c.csf_mask_erosion_mm
+            seg_preproc.inputs.wm_mask_erosion_mm.wm_mask_erosion_mm=c.wm_mask_erosion_mm
+            seg_preproc.inputs.gm_mask_erosion_mm.gm_mask_erosion_mm=c.gm_mask_erosion_mm
+
+            seg_preproc.inputs.csf_erosion_mm.csf_erosion_mm=c.csf_erosion_mm
+            seg_preproc.inputs.wm_erosion_mm.wm_erosion_mm=c.wm_erosion_mm
+            seg_preproc.inputs.gm_erosion_mm.gm_erosion_mm=c.gm_erosion_mm
+
+            node, out_file = strat['anatomical_brain_mask']
+            workflow.connect(node, out_file, seg_preproc, 'inputspec.brain_mask')
+
+            # TODO ASH review
+            if seg_preproc is None:
+                continue
+
+            node, out_file = strat['anatomical_brain']
+            workflow.connect(node, out_file,
+                             seg_preproc, 'inputspec.brain')
+
+            if strat.get('registration_method') == 'FSL':
+                node, out_file = strat['mni_to_anatomical_linear_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc,
+                                 'inputspec.standard2highres_mat')
+
+            elif strat.get('registration_method') == 'ANTS':
+                node, out_file = strat['ants_initial_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc,
+                                 'inputspec.standard2highres_init')
+
+                node, out_file = strat['ants_rigid_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc,
+                                 'inputspec.standard2highres_rig')
+
+                node, out_file = strat['ants_affine_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc,
+                                 'inputspec.standard2highres_mat')
+
+            workflow.connect(c.PRIORS_CSF, 'local_path',
+                                seg_preproc, 'inputspec.PRIOR_CSF')
+
+            workflow.connect(c.PRIORS_GRAY, 'local_path',
+                                seg_preproc, 'inputspec.PRIOR_GRAY')
+
+            workflow.connect(c.PRIORS_WHITE, 'local_path',
+                                seg_preproc, 'inputspec.PRIOR_WHITE')
+
+            # TODO ASH review with forking function
+            if 0 in c.runSegmentationPreprocessing:
+                strat = strat.fork()
+                new_strat_list.append(strat)
+
+            if c.brain_use_erosion:
+                ero_imports = ['import scipy.ndimage as nd' , 'import numpy as np', 'import nibabel as nb', 'import os']
+                eroded_mask = pe.Node(util.Function(input_names = ['roi_mask', 'skullstrip_mask', 'mask_erosion_mm', 'mask_erosion_prop'],
+                                                    output_names = ['output_roi_mask', 'eroded_skullstrip_mask'],
+                                                    function = mask_erosion,
+                                                    imports = ero_imports),
+                                                    name='erode_skullstrip_brain_mask')
+                eroded_mask.inputs.mask_erosion_mm = c.brain_mask_erosion_mm
+                workflow.connect(anat_preproc, 'outputspec.brain_mask', eroded_mask, 'skullstrip_mask')
+                workflow.connect(seg_preproc, 'outputspec.csf_probability_map', eroded_mask, 'roi_mask')
+
+                strat.update_resource_pool({'anatomical_eroded_brain_mask': (eroded_mask, 'eroded_skullstrip_mask')})
+
+            strat.append_name(seg_preproc.name)
+            strat.update_resource_pool({
+                'anatomical_gm_mask': (seg_preproc, 'outputspec.gm_mask'),
+                'anatomical_csf_mask': (seg_preproc, 'outputspec.csf_mask'),
+                'anatomical_wm_mask': (seg_preproc, 'outputspec.wm_mask'),
+                'seg_probability_maps': (seg_preproc, 'outputspec.probability_maps'),
+                'seg_mixeltype': (seg_preproc, 'outputspec.mixeltype'),
+                'seg_partial_volume_map': (seg_preproc, 'outputspec.partial_volume_map'),
+                'seg_partial_volume_files': (seg_preproc, 'outputspec.partial_volume_files')
+            })
+
+    strat_list += new_strat_list
+
+    if 'T1_template' in c.template_based_segmentation:
+
+        for num_strat, strat in enumerate(strat_list):
+
+            nodes = strat.get_nodes_names()
+
+            if not any(o in c.template_based_segmentation for o in ['EPI_template', 'T1_template', 'None']):
+                err = '\n\n[!] C-PAC says: Your template based segmentation ' \
+                    'setting does not include either \'EPI_template\' or \'T1_template\'.\n\n' \
+                    'Options you provided:\ntemplate_based_segmentation: {0}' \
+                    '\n\n'.format(str(c.template_based_segmentation))
+                raise Exception(err)
+
+            if strat.get('registration_method') == 'FSL':
+                use_ants = False
+            elif strat.get('registration_method') == 'ANTS':
+                use_ants = True
+
+            seg_preproc_template_based = create_seg_preproc_template_based(use_ants=use_ants,
+                                                             wf_name='seg_preproc_t1_template_{0}'.format(num_strat))
+
+            if seg_preproc_template_based is None:
+                continue
+
+            node, out_file = strat['anatomical_brain']
+            workflow.connect(node, out_file,
+                             seg_preproc_template_based, 'inputspec.brain')
+
+            if strat.get('registration_method') == 'FSL':
+                node, out_file = strat['mni_to_anatomical_linear_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc_template_based,
+                                 'inputspec.standard2highres_mat')
+
+            elif strat.get('registration_method') == 'ANTS':
+                node, out_file = strat['ants_initial_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc_template_based,
+                                 'inputspec.standard2highres_init')
+
+                node, out_file = strat['ants_rigid_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc_template_based,
+                                 'inputspec.standard2highres_rig')
+
+                node, out_file = strat['ants_affine_xfm']
+                workflow.connect(node, out_file,
+                                 seg_preproc_template_based,
+                                 'inputspec.standard2highres_mat')
+
+            workflow.connect(c.template_based_segmentation_CSF, 'local_path',
+                                seg_preproc_template_based, 'inputspec.CSF_template')
+
+            workflow.connect(c.template_based_segmentation_GRAY, 'local_path',
+                                seg_preproc_template_based, 'inputspec.GRAY_template')
+
+            workflow.connect(c.template_based_segmentation_WHITE, 'local_path',
+                                seg_preproc_template_based, 'inputspec.WHITE_template')
+
+            if 'None' in c.template_based_segmentation:
+                strat = strat.fork()
+                new_strat_list.append(strat)
+
+            strat.append_name(seg_preproc_template_based.name)
+            strat.update_resource_pool({
+                'anatomical_gm_mask': (seg_preproc_template_based, 'outputspec.gm_mask'),
+                'anatomical_csf_mask': (seg_preproc_template_based, 'outputspec.csf_mask'),
+                'anatomical_wm_mask': (seg_preproc_template_based, 'outputspec.wm_mask')
+            })
+
+    strat_list += new_strat_list
+
+    return workflow, strat_list
